requirements.txt
ddt==1.3.1
django==2.2.12
django-cors-headers==3.2.1
drf-yasg==1.17.1
django-simple-history==2.10.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
pymongo
psycopg2-binary
gunicorn==20.0.4
